backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
easyocr
pillow
python-dotenv
